+++ CountlyEvent.m
@@ -9,6 +9,9 @@
 @implementation CountlyEvent
 
 NSString* const kCountlyEventKeyKey           = @"key";
+NSString* const kCountlyEventKeyID            = @"id";
+NSString* const kCountlyEventKeyCVID          = @"cvid";
+NSString* const kCountlyEventKeyPVID          = @"pvid";
 NSString* const kCountlyEventKeySegmentation  = @"segmentation";
 NSString* const kCountlyEventKeyCount         = @"count";
 NSString* const kCountlyEventKeySum           = @"sum";
@@ -25,6 +28,9 @@ NSString* const kCountlyEventKeyDuration      = @"dur";
     {
         eventData[kCountlyEventKeySegmentation] = self.segmentation;
     }
+    eventData[kCountlyEventKeyID] = self.ID;
+    eventData[kCountlyEventKeyCVID] = self.CVID;
+    eventData[kCountlyEventKeyPVID] = self.PVID;
     eventData[kCountlyEventKeyCount] = @(self.count);
     eventData[kCountlyEventKeySum] = @(self.sum);
     eventData[kCountlyEventKeyTimestamp] = @((long long)(self.timestamp * 1000));
@@ -39,6 +45,9 @@ NSString* const kCountlyEventKeyDuration      = @"dur";
     if (self = [super init])
     {
         self.key = [decoder decodeObjectForKey:NSStringFromSelector(@selector(key))];
+        self.ID = [decoder decodeObjectForKey:NSStringFromSelector(@selector(ID))];
+        self.CVID = [decoder decodeObjectForKey:NSStringFromSelector(@selector(CVID))];
+        self.PVID = [decoder decodeObjectForKey:NSStringFromSelector(@selector(PVID))];
         self.segmentation = [decoder decodeObjectForKey:NSStringFromSelector(@selector(segmentation))];
         self.count = [decoder decodeIntegerForKey:NSStringFromSelector(@selector(count))];
         self.sum = [decoder decodeDoubleForKey:NSStringFromSelector(@selector(sum))];
@@ -54,6 +63,9 @@ NSString* const kCountlyEventKeyDuration      = @"dur";
 - (void)encodeWithCoder:(NSCoder *)encoder
 {
     [encoder encodeObject:self.key forKey:NSStringFromSelector(@selector(key))];
+    [encoder encodeObject:self.ID forKey:NSStringFromSelector(@selector(ID))];
+    [encoder encodeObject:self.CVID forKey:NSStringFromSelector(@selector(CVID))];
+    [encoder encodeObject:self.PVID forKey:NSStringFromSelector(@selector(PVID))];
     [encoder encodeObject:self.segmentation forKey:NSStringFromSelector(@selector(segmentation))];
     [encoder encodeInteger:self.count forKey:NSStringFromSelector(@selector(count))];
     [encoder encodeDouble:self.sum forKey:NSStringFromSelector(@selector(sum))];